refactor: consolidate naming of services in `workbench-application.core` module
