+++ projects/scion/workbench-application.core/src/lib/activity.service.ts
@@ -9,7 +9,7 @@
  */
 
 import { Disposable, Service } from './metadata';
-import { MessageBus } from './message-bus-service';
+import { MessageBus } from './message-bus.service';
 import { Observable, ReplaySubject, Subject } from 'rxjs';
 import { filter, takeUntil } from 'rxjs/operators';
 import { UUID } from './uuid.util';
