+++ projects/scion/workbench-application.core/src/lib/message-bus.service.ts
