+++ projects/scion/workbench-application.core/src/lib/platform-event-bus.service.ts
