+++ projects/scion/workbench-application.core/src/lib/popup.service.ts
@@ -8,7 +8,7 @@
  *  SPDX-License-Identifier: EPL-2.0
  */
 
-import { MessageBus } from './message-bus-service';
+import { MessageBus } from './message-bus.service';
 import { Service } from './metadata';
 import { Platform } from './platform';
 import { Params } from '@angular/router';
